+++ src/2_Remote.h
@@ -2,11 +2,13 @@
 
 // Select (remove //) the remote configuration profile you have:
 // #define FLYSKY_FS_I6X            // <------- Flysky FS-i6x
-// #define FLYSKY_FS_I6S                // <------- Flysky FS-i6s
+// #define FLYSKY_FS_I6S            // <------- Flysky FS-i6s
 // #define FLYSKY_FS_I6S_LOADER     // <------- Flysky FS-i6s for BURNIE222 Volvo L120H loader (use IBUS communication setting)
-// #define FLYSKY_FS_I6S_EXCAVATOR // <------- Flysky FS-i6s for KABOLITE K336 hydraulic excavator (use IBUS communication setting)
-// #define FRSKY_TANDEM_EXCAVATOR // <------- Frsky Tandem XE for hydraulic excavator (use SBUS communication setting)
- #define FRSKY_TANDEM_HARMONY_LOADER // <------- Frsky Tandem XE for Lukas Cajkar Harmony 370 (use SBUS communication setting)
+// #define FLYSKY_FS_I6S_EXCAVATOR  // <------- Flysky FS-i6s for KABOLITE K336 hydraulic excavator (use IBUS communication setting)
+// #define FRSKY_TANDEM_EXCAVATOR   // <------- Frsky Tandem XE for hydraulic excavator (use SBUS communication setting)
+// #define FRSKY_TANDEM_HARMONY_LOADER // <------- Frsky Tandem XE for Lukas Cajkar Harmony 370 (use SBUS communication setting)
+// #define FRSKY_TANDEM_EXCAVATOR  // <------- Frsky Tandem XE for hydraulic excavator (use SBUS communication setting)
+ #define FRSKY_TANDEM_CRANE      // <------- Frsky Tandem XE for Mushroom3D rough terrain crane (use SBUS communication setting)
 // #define FLYSKY_GT5              // <------- Flysky GT5 / Reely GT6 EVO / Absima CR6P
 // #define RGT_EX86100             // <------- MT-305 remote delivered with RGT EX86100 crawler (use PWM communication setting)
 // #define GRAUPNER_MZ_12          // <------- Graupner MZ-12 PRO
@@ -26,8 +28,8 @@
 // PWM servo signal communication (CH1 - CH6 headers, 6 channels) --------
 // PWM mode active, if SBUS, IBUS, SUMD and PPM are disabled (// in front of #define)
 
-// SBUS communication (RX header, 13 channels. This is my preferred communication protocol)--------
-#define SBUS_COMMUNICATION // control signals are coming in via the SBUS interface (comment it out for classic PWM RC signals)
+// SBUS communication (RX header, 16 channels. This is my preferred communication protocol)--------
+ #define SBUS_COMMUNICATION // control signals are coming in via the SBUS interface (comment it out for classic PWM RC signals)
 // NOTE: "boolean sbusInverted = true / false" was moved to the remote configuration profiles, so you don't have to change it
 uint32_t sbusBaud = 100000;         // Standard is 100000. Try to lower it, if your channels are coming in unstable. Working range is about 96000 - 104000.
 #define EMBEDDED_SBUS               // Embedded SBUS code is used instead of SBUS library, if defined (recommended, don't change it)
@@ -105,9 +107,12 @@ uint16_t sbusFailsafeTimeout = 100; // Failsafe is triggered after this timeout 
 #define HAZARDS NONE         // CH11
 #define INDICATOR_LEFT NONE  // CH12
 #define INDICATOR_RIGHT NONE // CH13
+#define CH_14 14             // CH14
+#define CH_15 15             // CH15
+#define CH_16 16             // CH16
 
 // Channels reversed or not
-boolean channelReversed[14] = {
+boolean channelReversed[17] = {
     false, // CH0 (unused)
     false, // CH1
     false, // CH2
@@ -121,11 +126,14 @@ boolean channelReversed[14] = {
     false, // CH10
     false, // CH11
     false, // CH12
-    false  // CH13
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
 };
 
 // Channels auto zero adjustment or not (don't use it for channels without spring centered neutral position, switches or unused channels)
-boolean channelAutoZero[14] = {
+boolean channelAutoZero[17] = {
     false, // CH0 (unused)
     true,  // CH1
     false, // CH2
@@ -139,7 +147,10 @@ boolean channelAutoZero[14] = {
     false, // CH10
     false, // CH11
     false, // CH12
-    false  // CH13
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
 };
 
 // Channels signal range calibration -----
@@ -174,9 +185,12 @@ boolean sbusInverted = true; // true = standard (non inverted) SBUS signal
 #define HAZARDS NONE         // CH11
 #define INDICATOR_LEFT NONE  // CH12
 #define INDICATOR_RIGHT NONE // CH13
+#define CH_14 14             // CH14
+#define CH_15 15             // CH15
+#define CH_16 16             // CH16
 
 // Channels reversed or not
-boolean channelReversed[14] = {
+boolean channelReversed[17] = {
     false, // CH0 (unused)
     false, // CH1
     false, // CH2
@@ -190,11 +204,14 @@ boolean channelReversed[14] = {
     false, // CH10
     false, // CH11
     false, // CH12
-    false  // CH13
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
 };
 
 // Channels auto zero adjustment or not (don't use it for channels without spring centered neutral position, switches or unused channels)
-boolean channelAutoZero[14] = {
+boolean channelAutoZero[17] = {
     false, // CH0 (unused)
     true,  // CH1
     false, // CH2
@@ -208,7 +225,10 @@ boolean channelAutoZero[14] = {
     false, // CH10
     false, // CH11
     false, // CH12
-    false  // CH13
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
 };
 
 // Channels signal range calibration -----
@@ -247,9 +267,12 @@ boolean sbusInverted = true; // true = standard (non inverted) SBUS signal
 #define HAZARDS NONE         // CH11
 #define INDICATOR_LEFT NONE  // CH12
 #define INDICATOR_RIGHT NONE // CH13
+#define CH_14 14             // CH14
+#define CH_15 15             // CH15
+#define CH_16 16             // CH16
 
 // Channels reversed or not
-boolean channelReversed[14] = {
+boolean channelReversed[17] = {
     false, // CH0 (unused)
     false, // CH1
     false, // CH2
@@ -263,11 +286,14 @@ boolean channelReversed[14] = {
     false, // CH10
     false, // CH11
     false, // CH12
-    false  // CH13
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
 };
 
 // Channels auto zero adjustment or not (don't use it for channels without spring centered neutral position, switches or unused channels)
-boolean channelAutoZero[14] = {
+boolean channelAutoZero[17] = {
     false, // CH0 (unused)
     true,  // CH1
     false, // CH2
@@ -281,7 +307,10 @@ boolean channelAutoZero[14] = {
     false, // CH10
     false, // CH11
     false, // CH12
-    false  // CH13
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
 };
 
 // Channels signal range calibration -----
@@ -318,9 +347,12 @@ boolean sbusInverted = true; // true = standard (non inverted) SBUS signal
 #define HAZARDS NONE         // CH11
 #define INDICATOR_LEFT NONE  // CH12
 #define INDICATOR_RIGHT NONE // CH13
+#define CH_14 14             // CH14
+#define CH_15 15             // CH15
+#define CH_16 16             // CH16
 
 // Channels reversed or not
-boolean channelReversed[14] = {
+boolean channelReversed[17] = {
     false, // CH0 (unused)
     false, // CH1
     false, // CH2
@@ -334,11 +366,14 @@ boolean channelReversed[14] = {
     false, // CH10
     false, // CH11
     false, // CH12
-    false  // CH13
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
 };
 
 // Channels auto zero adjustment or not (don't use it for channels without spring centered neutral position, switches or unused channels)
-boolean channelAutoZero[14] = {
+boolean channelAutoZero[17] = {
     false, // CH0 (unused)
     true,  // CH1
     true,  // CH2
@@ -352,7 +387,10 @@ boolean channelAutoZero[14] = {
     false, // CH10
     false, // CH11
     false, // CH12
-    false  // CH13
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
 };
 
 // Channels signal range calibration -----
@@ -389,9 +427,12 @@ boolean sbusInverted = true; // true = standard (non inverted) SBUS signal
 #define HAZARDS NONE         // CH11
 #define INDICATOR_LEFT NONE  // CH12
 #define INDICATOR_RIGHT NONE // CH13
+#define CH_14 14             // CH14
+#define CH_15 15             // CH15
+#define CH_16 16             // CH16
 
 // Channels reversed or not
-boolean channelReversed[14] = {
+boolean channelReversed[17] = {
     false, // CH0 (unused)
     false, // CH1
     false, // CH2
@@ -405,11 +446,14 @@ boolean channelReversed[14] = {
     false, // CH10
     false, // CH11
     false, // CH12
-    false  // CH13
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
 };
 
 // Channels auto zero adjustment or not (don't use it for channels without spring centered neutral position, switches or unused channels)
-boolean channelAutoZero[14] = {
+boolean channelAutoZero[17] = {
     false, // CH0 (unused)
     true,  // CH1
     true,  // CH2
@@ -423,7 +467,10 @@ boolean channelAutoZero[14] = {
     false, // CH10
     false, // CH11
     false, // CH12
-    false  // CH13
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
 };
 
 // Channels signal range calibration -----
@@ -460,9 +507,12 @@ boolean sbusInverted = true; // true = standard (non inverted) SBUS signal
 #define HAZARDS NONE         // CH11
 #define INDICATOR_LEFT NONE  // CH12
 #define INDICATOR_RIGHT NONE // CH13
+#define CH_14 14             // CH14
+#define CH_15 15             // CH15
+#define CH_16 16             // CH16
 
 // Channels reversed or not
-boolean channelReversed[14] = {
+boolean channelReversed[17] = {
     false, // CH0 (unused)
     true, // CH1
     true, // CH2
@@ -476,11 +526,14 @@ boolean channelReversed[14] = {
     false, // CH10
     false, // CH11
     false, // CH12
-    false  // CH13
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
 };
 
 // Channels auto zero adjustment or not (don't use it for channels without spring centered neutral position, switches or unused channels)
-boolean channelAutoZero[14] = {
+boolean channelAutoZero[17] = {
     false, // CH0 (unused)
     true,  // CH1
     true, // CH2
@@ -494,12 +547,95 @@ boolean channelAutoZero[14] = {
     false, // CH10
     false, // CH11
     false, // CH12
-    false  // CH13
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
 };
 
 // Channels signal range calibration -----
 const uint16_t pulseNeutral = 30;
 const uint16_t pulseSpan = 495;
+
+// Automatic or manual modes -----
+// #define AUTO_LIGHTS
+// #define AUTO_ENGINE_ON_OFF
+// #define AUTO_INDICATORS
+
+// SBUS mode ----
+boolean sbusInverted = true; // true = standard (non inverted) SBUS signal
+
+#endif
+
+// Frsky Tandem XE remote configuration profile (for excavators only) ---------------------------------------------------------------------------------
+#ifdef FRSKY_TANDEM_CRANE
+
+// NOTE: The vehicle file needs to contain #define EXCAVATOR_MODE
+
+// Channel assignment (use NONE for non existing channels!)
+// Remote channel #######   // Sound controller channel (pulseWidth[x]) ##########################################
+#define STEERING 4           // CH1 boom lift
+#define GEARBOX 5            // CH2 boom extension
+#define THROTTLE 1           // CH3 throttle & brake
+#define HORN 10              // CH4 horn and siren (slider right)
+#define FUNCTION_R 11        // CH5 jake brake, high / low beam, headlight flasher, engine on / off SFR-Source1
+#define FUNCTION_L 12        // CH6 rotating beacon, hazards SFR-Source2
+#define POT2 6               // CH7 main rope
+#define MODE1 8              // CH8 swing
+#define MODE2 7              // CH9 fast rope
+#define MOMENTARY1 NONE      // CH10
+#define HAZARDS NONE         // CH11
+#define INDICATOR_LEFT 9     // CH12 Outrigger booms (all together)
+#define INDICATOR_RIGHT 13   // CH13 Support cylinder front left
+#define CH_14 14             // CH14 Support cylinder front right
+#define CH_15 15             // CH15 Support cylinder rear left
+#define CH_16 16             // CH16 Support cylinder rear right
+
+// Channels reversed or not
+boolean channelReversed[17] = {
+    false, // CH0 (unused)
+    true, // CH1
+    false, // CH2
+    false, // CH3
+    false, // CH4
+    true, // CH5
+    false, // CH6
+    false, // CH7
+    false, // CH8
+    false, // CH9
+    false, // CH10
+    false, // CH11
+    false, // CH12
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
+};
+
+// Channels auto zero adjustment or not (don't use it for channels without spring centered neutral position, switches or unused channels)
+boolean channelAutoZero[17] = {
+    false, // CH0 (unused)
+    false,  // CH1
+    false,  // CH2
+    true, // CH3
+    false, // CH4
+    false,  // CH5
+    false,  // CH6
+    false,  // CH7
+    false,  // CH8
+    false, // CH9
+    false, // CH10
+    false, // CH11
+    false, // CH12
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
+};
+
+// Channels signal range calibration -----
+const uint16_t pulseNeutral = 30;
+const uint16_t pulseSpan = 480;
 
 // Automatic or manual modes -----
 // #define AUTO_LIGHTS
@@ -531,9 +667,12 @@ boolean sbusInverted = true; // true = standard (non inverted) SBUS signal
 #define HAZARDS NONE         // CH11
 #define INDICATOR_LEFT NONE  // CH12
 #define INDICATOR_RIGHT NONE // CH13
+#define CH_14 14             // CH14
+#define CH_15 15             // CH15
+#define CH_16 16             // CH16
 
 // Channels reversed or not
-boolean channelReversed[14] = {
+boolean channelReversed[17] = {
     false, // CH0 (unused)
     false, // CH1
     false, // CH2
@@ -547,11 +686,14 @@ boolean channelReversed[14] = {
     false, // CH10
     false, // CH11
     false, // CH12
-    false  // CH13
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
 };
 
 // Channels auto zero adjustment or not (don't use it for channels without spring centered neutral position, switches or unused channels)
-boolean channelAutoZero[14] = {
+boolean channelAutoZero[17] = {
     false, // CH0 (unused)
     true,  // CH1
     true,  // CH2
@@ -565,7 +707,10 @@ boolean channelAutoZero[14] = {
     false, // CH10
     false, // CH11
     false, // CH12
-    false  // CH13
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
 };
 
 // Channels signal range calibration -----
@@ -626,9 +771,12 @@ boolean sbusInverted = true; // true = standard (non inverted) SBUS signal
 #define HAZARDS 3            // CH11
 #define INDICATOR_LEFT NONE  // CH12
 #define INDICATOR_RIGHT NONE // CH13
+#define CH_14 14             // CH14
+#define CH_15 15             // CH15
+#define CH_16 16             // CH16
 
 // Channels reversed or not
-boolean channelReversed[14] = {
+boolean channelReversed[17] = {
     false, // CH0 (unused)
     false, // CH1
     false, // CH2
@@ -642,11 +790,14 @@ boolean channelReversed[14] = {
     false, // CH10
     false, // CH11
     false, // CH12
-    false  // CH13
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
 };
 
 // Channels auto zero adjustment or not (don't use it for channels without spring centered neutral position, switches or unused channels)
-boolean channelAutoZero[14] = {
+boolean channelAutoZero[17] = {
     false, // CH0 (unused)
     true,  // CH1
     false, // CH2
@@ -660,7 +811,10 @@ boolean channelAutoZero[14] = {
     false, // CH10
     false, // CH11
     false, // CH12
-    false  // CH13
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
 };
 
 // Channels signal range calibration -----
@@ -716,9 +870,12 @@ boolean sbusInverted = true; // true = standard (non inverted) SBUS signal
 #define HAZARDS NONE         // CH11
 #define INDICATOR_LEFT NONE  // CH12
 #define INDICATOR_RIGHT NONE // CH13
+#define CH_14 14             // CH14
+#define CH_15 15             // CH15
+#define CH_16 16             // CH16
 
 // Channels reversed or not
-boolean channelReversed[14] = {
+boolean channelReversed[17] = {
     false, // CH0 (unused)
     false, // CH1
     false, // CH2
@@ -732,11 +889,14 @@ boolean channelReversed[14] = {
     false, // CH10
     false, // CH11
     false, // CH12
-    false  // CH13
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
 };
 
 // Channels auto zero adjustment or not (don't use it for channels without spring centered neutral position, switches or unused channels)
-boolean channelAutoZero[14] = {
+boolean channelAutoZero[17] = {
     false, // CH0 (unused)
     true,  // CH1
     false, // CH2
@@ -750,7 +910,10 @@ boolean channelAutoZero[14] = {
     false, // CH10
     false, // CH11
     false, // CH12
-    false  // CH13
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
 };
 
 // Channels signal range calibration -----
@@ -785,9 +948,12 @@ boolean sbusInverted = true; // true = standard (non inverted) SBUS signal
 #define HAZARDS 9            // CH11
 #define INDICATOR_LEFT NONE  // CH12
 #define INDICATOR_RIGHT NONE // CH13
+#define CH_14 14             // CH14
+#define CH_15 15             // CH15
+#define CH_16 16             // CH16
 
 // Channels reversed or not
-boolean channelReversed[14] = {
+boolean channelReversed[17] = {
     false, // CH0 (unused)
     false, // CH1
     false, // CH2
@@ -801,11 +967,14 @@ boolean channelReversed[14] = {
     false, // CH10
     false, // CH11
     false, // CH12
-    false  // CH13
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
 };
 
 // Channels auto zero adjustment or not (don't use it for channels without spring centered neutral position, switches or unused channels)
-boolean channelAutoZero[14] = {
+boolean channelAutoZero[17] = {
     false, // CH0 (unused)
     false, // CH1
     false, // CH2
@@ -819,7 +988,10 @@ boolean channelAutoZero[14] = {
     false, // CH10
     false, // CH11
     false, // CH12
-    false  // CH13
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
 };
 
 // Channels signal range calibration -----
@@ -841,22 +1013,25 @@ boolean sbusInverted = true; // true = standard (non inverted) SBUS signal
 
 // Channel assignment (use NONE for non existing channels!)
 // Remote channel #######   // Sound controller channel ##########################################
-#define STEERING 1         // CH1 steering
-#define GEARBOX 2          // CH2 3 position switch for gearbox (left throttle in tracked mode)
-#define THROTTLE 3         // CH3 throttle & brake (right throttle in tracked mode)
-#define HORN 5             // CH4 horn and bluelight / siren
-#define FUNCTION_R 4       // CH5 jake brake, high / low beam, headlight flasher, engine on / off
-#define FUNCTION_L NONE    // CH6 indicators, hazards
-#define POT2 16            // CH7 pot2
-#define MODE1 6            // CH8 mode 1 switch
-#define MODE2 7            // CH9 mode 2 switch
-#define MOMENTARY1 8       // CH10
-#define HAZARDS 9          // CH11
-#define INDICATOR_LEFT 10  // CH12
-#define INDICATOR_RIGHT 11 // CH13
+#define STEERING 1           // CH1 steering
+#define GEARBOX 2            // CH2 3 position switch for gearbox (left throttle in tracked mode)
+#define THROTTLE 3           // CH3 throttle & brake (right throttle in tracked mode)
+#define HORN 5               // CH4 horn and bluelight / siren
+#define FUNCTION_R 4         // CH5 jake brake, high / low beam, headlight flasher, engine on / off
+#define FUNCTION_L NONE      // CH6 indicators, hazards
+#define POT2 16              // CH7 pot2
+#define MODE1 6              // CH8 mode 1 switch
+#define MODE2 7              // CH9 mode 2 switch
+#define MOMENTARY1 8         // CH10
+#define HAZARDS 9            // CH11
+#define INDICATOR_LEFT 10    // CH12
+#define INDICATOR_RIGHT 11   // CH13
+#define CH_14 14             // CH14
+#define CH_15 15             // CH15
+#define CH_16 16             // CH16
 
 // Channels reversed or not
-boolean channelReversed[14] = {
+boolean channelReversed[17] = {
     false, // CH0 (unused)
     false, // CH1
     false, // CH2
@@ -870,11 +1045,14 @@ boolean channelReversed[14] = {
     false, // CH10
     false, // CH11
     false, // CH12
-    false  // CH13
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
 };
 
 // Channels auto zero adjustment or not (don't use it for channels without spring centered neutral position, switches or unused channels)
-boolean channelAutoZero[14] = {
+boolean channelAutoZero[17] = {
     false, // CH0 (unused)
     true,  // CH1 true
     false, // CH2
@@ -888,7 +1066,10 @@ boolean channelAutoZero[14] = {
     false, // CH10
     false, // CH11
     false, // CH12
-    false  // CH13
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
 };
 
 // Channels signal range calibration -----
@@ -910,22 +1091,25 @@ boolean sbusInverted = false; // false = non standard (inverted) SBUS signal
 
 // Channel assignment (use NONE for non existing channels!)
 // Remote channel #######   // Sound controller channel ##########################################
-#define STEERING 1         // CH1 steering
-#define GEARBOX NONE       // CH2 3 position switch for gearbox (left throttle in tracked mode)
-#define THROTTLE 3         // CH3 throttle & brake (right throttle in tracked mode)
-#define HORN 5             // CH4 horn and bluelight / siren
-#define FUNCTION_R NONE    // CH5 jake brake, high / low beam, headlight flasher, engine on / off
-#define FUNCTION_L NONE    // CH6 indicators, hazards
-#define POT2 16            // CH7 pot2
-#define MODE1 6            // CH8 mode 1 switch
-#define MODE2 7            // CH9 mode 2 switch
-#define MOMENTARY1 8       // CH10
-#define HAZARDS 9          // CH11
-#define INDICATOR_LEFT 10  // CH12
-#define INDICATOR_RIGHT 11 // CH13
+#define STEERING 1           // CH1 steering
+#define GEARBOX NONE         // CH2 3 position switch for gearbox (left throttle in tracked mode)
+#define THROTTLE 3           // CH3 throttle & brake (right throttle in tracked mode)
+#define HORN 5               // CH4 horn and bluelight / siren
+#define FUNCTION_R NONE      // CH5 jake brake, high / low beam, headlight flasher, engine on / off
+#define FUNCTION_L NONE      // CH6 indicators, hazards
+#define POT2 16              // CH7 pot2
+#define MODE1 6              // CH8 mode 1 switch
+#define MODE2 7              // CH9 mode 2 switch
+#define MOMENTARY1 8         // CH10
+#define HAZARDS 9            // CH11
+#define INDICATOR_LEFT 10    // CH12
+#define INDICATOR_RIGHT 11   // CH13
+#define CH_14 14             // CH14
+#define CH_15 15             // CH15
+#define CH_16 16             // CH16
 
 // Channels reversed or not
-boolean channelReversed[14] = {
+boolean channelReversed[17] = {
     false, // CH0 (unused)
     false, // CH1
     false, // CH2
@@ -939,11 +1123,14 @@ boolean channelReversed[14] = {
     false, // CH10
     false, // CH11
     false, // CH12
-    false  // CH13
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
 };
 
 // Channels auto zero adjustment or not (don't use it for channels without spring centered neutral position, switches or unused channels)
-boolean channelAutoZero[14] = {
+boolean channelAutoZero[17] = {
     false, // CH0 (unused)
     true,  // CH1 true
     false, // CH2
@@ -957,7 +1144,10 @@ boolean channelAutoZero[14] = {
     false, // CH10
     false, // CH11
     false, // CH12
-    false  // CH13
+    false, // CH13
+    false, // CH14
+    false, // CH15
+    false  // CH16
 };
 
 // Channels signal range calibration -----