+++ src/4_Transmission.h
@@ -30,7 +30,7 @@ uint16_t automaticReverseAccelerationPercentage = 100;
 uint16_t lowRangePercentage = 58; // WPL 2 speed ratios = 29:1, 17:1 = 58% in low range. You may want to change this for other 2 speed transmissions
 
 // Transmission controls options ===========================================================================================================
-// #define SEMI_AUTOMATIC This will simulate a semi automatic transmission. Shifting is not controlled by the 3 position switch in this mode!
+ #define SEMI_AUTOMATIC // This will simulate a semi automatic transmission. Shifting is not controlled by the 3 position switch in this mode!
 // #define SEMI_AUTOMATIC // Works for VIRTUAL_3_SPEED or real 3 speed transmission. Don't select this @ the same time as VIRTUAL_16_SPEED_SEQUENTIAL
 
 // #define MODE1_SHIFTING The 2 speed transmission is shifted by the "Mode 1" button instead of the 3 position switch.